+++ tests/testthat/tests.restructure.R
@@ -131,9 +131,9 @@ test_that("change_old_sm return expected lists", {
             paste0("s_sx1kEQ9[", 1:9, "]"))
   dims <- list(sds_sigma_t2x0 = numeric(0), sds_sx1kEQ9 = numeric(0),
                s_sigma_t2x0 = 6, s_sx1kEQ9 = 9)
-  bterms <- parse_bf(bf(y ~ s(x1, k = 9), sigma ~ t2(x0)), 
-                     family = gaussian())
-  expect_equivalent(brms:::change_old_sm(bterms, pars, dims), target)
+  bterms <- parse_bf(bf(y ~ s(x1, k = 9), sigma ~ t2(x0)))
+  dat <- data.frame(y = rnorm(100), x1 = rnorm(100), x0 = rnorm(100))
+  expect_equivalent(brms:::change_old_sm(bterms, dat, pars, dims), target)
 })
 
 test_that("change_old_mo returns expected lists", {